feat(aStar): 实现 A* 寻路算法原型并更新相关功能
- 新增 A* 寻路算法实现，包括 aStar.hpp、aStar.cpp、graph.hpp 和 graph.cpp 文件
- 更新 main.cpp，添加 A* 算法测试用例
- 新增 ground.hpp 和 ground.cpp，用于处理地形数据
- 更新 reader.hpp 和 reader.cpp，支持读取双精度浮点数数据
- 修改 gen_map.py 和 jpg2csv.py，调整地图生成和转换流程
- 更新 problem_flow.md，添加 A* 算法说明文档

+++ include/csvReader/reader.hpp
@@ -13,14 +13,14 @@ public:
     bool readFromFile(const std::string& filename);
     
     // 获取读取到的数据
-    const std::vector<std::vector<int>>& getData() const;
+    const std::vector<std::vector<double>>& getData() const;
     
     // 获取数据的行数和列数
     size_t getRows() const;
     size_t getCols() const;
 
 private:
-    std::vector<std::vector<int>> data;
+    std::vector<std::vector<double>> data;
     size_t rows;
     size_t cols;
 };
